requirements.txt
amqp>=2.0.3
anyjson>=0.3.3
appnope>=0.1.0
autopep8>=1.2.4
Babel>=2.3.4
backports.shutil-get-terminal-size>=1.0.0
billiard>=3.5.0.0
celery>=3.1.23
click>=6.6
coverage>=4.2
decorator>=4.0.10
dj-database-url>=0.4.1
Django>=1.9
django-bootstrap3>=7.0.1
django-celery>=3.1.17
django-classy-tags>=0.8.0
django-cors-headers>=1.1.0
django-debug-toolbar>=1.5
django-filter>=0.14.0
django-formtools>=1.0
django-guardian>=1.4.6
django-phonenumber-field>=1.1.0
django-safedelete>=0.3.3
django-sekizai>=0.10.0
django-softdelete>=0.4
djangorestframework>=3.4.4
flake8>=3.0.4
gunicorn>=19.6.0
html5lib<=0.9999999
httplib2>=0.9.2
ipdb>=0.10.1
ipython>=5.1.0
ipython-genutils>=0.1.0
kombu>=3.0.35
mccabe>=0.5.2
newrelic>=2.68.0.50
pep8>=1.7.0
pexpect>=4.2.0
phonenumberslite>=7.5.2
pickleshare>=0.7.4
Pillow>=3.3.1
prompt-toolkit>=1.0.6
ptyprocess>=0.5.1
pycodestyle>=2.0.0
pyflakes>=1.2.3
Pygments>=2.1.3
PySocks>=1.5.7
python-dateutil>=2.5.3
pytz>=2016.6.1
redis>=2.10.5
requests>=2.11.1
requests-mock>=1.0.0
rq>=0.6.0
selenium>=2.53.6
simplegeneric>=0.8.1
six>=1.10.0
sqlparse>=0.2.1
traitlets>=4.2.2
twilio>=5.4.0
requests-mock>=1.0.0
django-filter>=0.14.0
psycopg2>=2.6.1
rq-scheduler==0.7.0
urllib3==1.18
braintree>=3.30.0
pyopenssl
django-session-security
django-axes
-e git+https://github.com/veox/python3-krakenex.git#egg=krakenx
-e git+https://github.com/byrnereese/uphold-sdk-python.git#egg=uphold